refactor: Move common components to a new directory structure and remove unused files
