+++ temporary/(canvas_js)/components/Header.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect, useRef } from 'react';
 import Link from 'next/link';
 import styles from '@/app/canvas/assets/Header.module.scss';
 import { LuPanelRightOpen, LuSave, LuCheck, LuX, LuPencil, LuFileText } from "react-icons/lu";
-import { getWorkflowName, saveWorkflowName } from '@/app/(common)/components/workflowStorage';
+import { getWorkflowName, saveWorkflowName } from '@/app/_common/components/workflowStorage';
 
 const Header = ({ onMenuClick, onSave, onLoad, onExport, onNewWorkflow, workflowName: externalWorkflowName, onWorkflowNameChange }) => {
     const [workflowName, setWorkflowName] = useState('Workflow');
