+++ temporary/(canvas_js)/components/SideMenuPanel/AddNodePanel.jsx
@@ -5,7 +5,7 @@ import NodeList from '@/app/canvas/components/Helper/NodeList';
 import DraggableNodeItem from '@/app/canvas/components/Helper/DraggableNodeItem';
 import { LuSearch, LuArrowLeft, LuBrainCircuit, LuShare2, LuWrench, LuX, LuRefreshCw } from 'react-icons/lu';
 import { SiLangchain } from "react-icons/si";
-import { useNodes } from '@/app/(common)/components/nodeHook';
+import { useNodes } from '@/app/_common/components/nodeHook';
 
 const iconMap = {
     LuBrainCircuit: <LuBrainCircuit />,
